i forgot to correct variable type in header

+++ src/mame/machine/nb1412m2.h
@@ -47,8 +47,8 @@ private:
 	uint16_t m_rom_address;
 	uint16_t m_adj_address;
 	uint16_t m_dac_start_address, m_dac_current_address;
-	int m_dac_frequency;
-	int m_timer_rate;
+	double m_dac_frequency;
+	uint8_t m_timer_rate;
 	uint8_t m_rom_op;
 	uint8_t m_const90;
 	bool m_timer_reg;
